patcher works for air client
Disabled Game client update as we dont pack it

+++ Sightstone.Patcher/Logic/Downloader.cs
@@ -1,8 +1,11 @@
-﻿using System;
+﻿using Ionic.Zlib;
+using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Net;
+using System.Threading;
+using System.Threading.Tasks;
 
 namespace Sightstone.Patcher.Logic
 {
@@ -11,7 +14,6 @@ namespace Sightstone.Patcher.Logic
     /// </summary>
     public class Downloader
     {
-        private List<DownloadFile> _downloading;
         private readonly List<DownloadFile> _finished = new List<DownloadFile>();
 
         public delegate void OnFinished();
@@ -20,78 +22,71 @@ namespace Sightstone.Patcher.Logic
         public delegate void OnProgressChanged(double downloaded, double toDownload);
         public event OnProgressChanged OnDownloadProgressChanged;
 
-        private double _downloadedBytes;
-        private double _bytesToDownload;
+        private long _downloadedBytes = 0;
+        private long _bytesToDownload;
+
+        private volatile int _webClientToCreate;
         public async void DownloadMultipleFiles(List<DownloadFile> filesToDownlad)
         {
+            _webClientToCreate = Client.MaximumWebClient;
             //_downloading = filesToDownlad;
+            _bytesToDownload = filesToDownlad.Sum(x => x.FileSize);
             foreach (var fileDlInfo in filesToDownlad)
             {
                 foreach (var paths in fileDlInfo.OutputPath.Where(paths => !Directory.Exists(Path.GetDirectoryName(paths))))
                     Directory.CreateDirectory(Path.GetDirectoryName(paths));
 
                 var dlInfo = fileDlInfo;
-                foreach (var overwrite in fileDlInfo.OutputPath.Where(overwrite => File.Exists(overwrite) && dlInfo.OverrideFiles))
+                
+                if (File.Exists(dlInfo.OutputPath[0]))
                 {
-                    try
+                    if (new FileInfo(dlInfo.OutputPath[0]).Length == dlInfo.FileSize)
                     {
-                        File.Delete(overwrite);
+                        Client.RunOnUIThread((() =>
+                        {
+                            var change = dlInfo.FileSize;
+                            _downloadedBytes = _downloadedBytes + change;
+                            OnDownloadProgressChanged?.Invoke(_downloadedBytes, _bytesToDownload);
+                        }));
                     }
-                    catch
-                    {
-                    }
-                }
-
-                var req = (HttpWebRequest)WebRequest.Create(fileDlInfo.DownloadUri);
-                req.Method = "HEAD";
-                using (var resp = (HttpWebResponse)req.GetResponse())
-                {
-                    _bytesToDownload = _bytesToDownload + resp.ContentLength;
-                    resp.Close();
+                    continue;
                 }
 
                 using (var client = new WebClient())
                 {
+                    while (_webClientToCreate == 0)
+                    {
+                        await Task.Delay(10);
+                    }
+                    _webClientToCreate--;
+
                     if (fileDlInfo.DownloadUri.ToString().EndsWith(".compressed"))
                     {
 
                         var info = fileDlInfo;
-                        var uncompressed = Ionic.Zlib.ZlibStream.UncompressBuffer(client.DownloadData(fileDlInfo.DownloadUri));
-                        try
+                        client.DownloadDataCompleted += (sender, e) =>
                         {
-                            foreach (var outputs in info.OutputPath)
-                                File.WriteAllBytes(outputs.Replace(".compressed", string.Empty), uncompressed);
-                        }
-                        catch
-                        {
-
-                        }
-                        //_downloading.Remove(info);
-                        _finished.Add(info);
-
-                        //if (_downloading.Count == 0 && _finished.Count != 0)
-                        //{
-                        //    OnFinishedDownloading?.Invoke();
-                        //}
+                            _webClientToCreate++;
+                            var uncompressed = ZlibStream.UncompressBuffer(e.Result);
+                            try
+                            {
+                                foreach (var outputs in info.OutputPath)
+                                    File.WriteAllBytes(outputs.Replace(".compressed", string.Empty), uncompressed);
+                            }
+                            catch
+                            {
+                            }
+                            _finished.Add(info);
+                            ((WebClient)sender).Dispose();
+                        };
+                        client.DownloadDataAsync(fileDlInfo.DownloadUri);
                     }
                     else
                     {
                         var info = fileDlInfo;
-                        client.DownloadFile(fileDlInfo.DownloadUri, fileDlInfo.OutputPath[0]);
-                        _finished.Add(info);
-
-                        if (info.OutputPath.Count() != 1)
+                        client.DownloadFileCompleted += (sender, e) =>
                         {
-                            foreach (var outputs in info.OutputPath.Skip(1))
-                            {
-                                File.Copy(info.OutputPath[0], outputs, info.OverrideFiles);
-                            }
-                        }
-                            client.DownloadFileCompleted += (sender, e) =>
-                        {
-                            //_downloading.Remove(info);
                             _finished.Add(info);
-
                             if (info.OutputPath.Count() != 1)
                             {
                                 foreach (var outputs in info.OutputPath.Skip(1))
@@ -99,20 +94,17 @@ namespace Sightstone.Patcher.Logic
                                     File.Copy(info.OutputPath[0], outputs, info.OverrideFiles);
                                 }
                             }
-                            //
-                            //if (_downloading.Count == 0 && _finished.Count != 0)
-                            //{
-                            //    OnFinishedDownloading?.Invoke();
-                            //}
+                            _webClientToCreate++;
                         };
+                        client.DownloadFileAsync(fileDlInfo.DownloadUri, fileDlInfo.OutputPath[0]);
                     }
 
-                    double lastbytes = 0;
+                    long lastbytes = 0;
                     client.DownloadProgressChanged += (sender, e) =>
                     {
                         Client.RunOnUIThread((() =>
                         {
-                            var bytesIn = double.Parse(e.BytesReceived.ToString());
+                            var bytesIn = e.BytesReceived;
                             var change = bytesIn - lastbytes;
                             _downloadedBytes = _downloadedBytes + change;
                             OnDownloadProgressChanged?.Invoke(_downloadedBytes, _bytesToDownload);
@@ -123,6 +115,8 @@ namespace Sightstone.Patcher.Logic
             }
             if (_finished.Count != 0)
             {
+                while (_webClientToCreate != Client.MaximumWebClient)
+                    await Task.Delay(10);
                 OnFinishedDownloading?.Invoke();
             }
         }
@@ -142,6 +136,11 @@ namespace Sightstone.Patcher.Logic
         /// The output path of the file
         /// </summary>
         public string[] OutputPath { get; set; }
+
+        /// <summary>
+        /// The size of the file
+        /// </summary>
+        public long FileSize { get; set; }
 
         /// <summary>
         /// If set to false, the if the file with the same name exists, the file will not be downloaded
